Update patch set 1
Patch Set 1:

Warning: Do not use this option on heterogeneous CPU configurations unless you have a capacity-aware RT scheduler. And, at this moment, NO CURRENTLY SHIPPING RT SCHEDULER IS CAPACITY-AWARE. We're working on one for EAS, but it's not yet available. The default RT scheduler is based purely on RT priorities and whether a CPU already has a RT thread of equal or higher priority. 

As a result, the default RT scheduler will happily move your relatively-long-running UI thread from a high-frequency big core to a little core at minimum frequency if a higher priority FIFO kthread happens to wake up on the same big core. This will introduce significant performance regressions. As this option has not yet been used on a shipping Android device, if you want to use it get in touch with the Android performance team to help you validate it.

Should we ship this?

Patch-set: 1
CC: Han Wang <16208@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>