apq8084-common: set ro.build.shutdown_timeout to 30
Users reported an unusual high battery consumption when the device was
actually supposed to be shut off.  Testing indicates that the device
gets shut down before rild has cleanly shut down, so it looks like the
radio component is in a limbo state after a shutdown and still needs
battery (and lots of it).

From the debug output it appears that rild needs 12 secs to shut down
radio for some reason.  Therefore set ro.build.shutdown_timeout to be on
the safe side.

Change-Id: I2e7585031636351accca0eab3441449eac8ddcbe
Signed-off-by: Corinna Vinschen <xda@vinschen.de>